+++ Solidity/IKWFundParticipant.sol
@@ -8,7 +8,7 @@ function notifyParticipant (bool giveup, uint128 summa_investors , uint128 summa
 function initialize (uint32 lock_time, 
                      uint32 unlock_time, 
                      uint128 quant,
-                     uint8 rate,
+                     uint16 rate,
                      uint32 kwf_lock_time) external ; 
 function sendFunds (TvmCell code) external; 
 function acknowledgeFunds () external ;
